save new structure prj
