--- app-cache/CMakeLists.txt
@@ -1,21 +1,0 @@
-cmake_minimum_required(VERSION 3.9)
-project(app_cache)
-
-set(CMAKE_CXX_STANDARD 20)
-
-configure_file(CMakeLists.txt.in
-        googletest-download/CMakeLists.txt)
-execute_process(COMMAND ${CMAKE_COMMAND} -G "${CMAKE_GENERATOR}" .
-        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/googletest-download )
-execute_process(COMMAND ${CMAKE_COMMAND} --build .
-        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/googletest-download )
-
-add_subdirectory(${CMAKE_BINARY_DIR}/googletest-src
-        ${CMAKE_BINARY_DIR}/googletest-build)
-
-include_directories("../${PROJECT_SOURCE_DIR}/info")
-
-enable_testing()
-add_subdirectory(tests)
-
-add_executable(app_cache main.cpp Cache.cpp Cache.h)
