--- app-cache/CMakeLists.txt.in
@@ -1,15 +1,0 @@
-cmake_minimum_required(VERSION 3.9)
-
-project(googletest-download NONE)
-
-include(ExternalProject)
-ExternalProject_Add(googletest
-        GIT_REPOSITORY    https://github.com/google/googletest.git
-        GIT_TAG           release-1.8.1
-        SOURCE_DIR        "${CMAKE_BINARY_DIR}/googletest-src"
-        BINARY_DIR        "${CMAKE_BINARY_DIR}/googletest-build"
-        CONFIGURE_COMMAND ""
-        BUILD_COMMAND     ""
-        INSTALL_COMMAND   ""
-        TEST_COMMAND      ""
-        )
