--- app-cache/info/Info.cpp
@@ -1,4 +1,0 @@
-//
-// Created by Sergei Alexeev on 11.04.2020.
-//
-
