--- app-network/client/Client.h
@@ -1,17 +1,0 @@
-//
-// Created by Sergei Alexeev on 11.04.2020.
-//
-
-#ifndef APP_NETWORK_CLIENT_H
-#define APP_NETWORK_CLIENT_H
-
-
-class Client {
-public:
-    bool start();
-
-private:
-};
-
-
-#endif //APP_NETWORK_CLIENT_H
