--- app-network/main.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-
-int main() {
-    std::cout << "Hello, World!" << std::endl;
-    return 0;
-}
