--- app-network/parser-json/Parser.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Sergei Alexeev on 11.04.2020.
-//
-
-#include "Parser.h"
