--- app-network/parser-json/Parser.h
@@ -1,14 +1,0 @@
-//
-// Created by Sergei Alexeev on 11.04.2020.
-//
-
-#ifndef APP_NETWORK_PARSER_H
-#define APP_NETWORK_PARSER_H
-
-
-class Parser {
-
-};
-
-
-#endif //APP_NETWORK_PARSER_H
