+++ network-cache/main.cpp
@@ -1,0 +1,12 @@
+#include <iostream>
+#include "network/AppNetwork.h"
+#include "cache/Cache.h"
+
+using namespace std;
+
+
+// main тут только для тестирования во время разработки, эта часть проекта должа быть либой
+int main() {
+    std::cout << "Hello, World!" << std::endl;
+    return 0;
+}
