+++ network-cache/network/AppNetwork.cpp
@@ -1,0 +1,5 @@
+//
+// Created by Sergei Alexeev on 13.04.2020.
+//
+
+#include "AppNetwork.h"
