+++ network-cache/network/AppNetwork.h
@@ -1,0 +1,14 @@
+//
+// Created by Sergei Alexeev on 13.04.2020.
+//
+
+#ifndef NETWORK_CACHE_APPNETWORK_H
+#define NETWORK_CACHE_APPNETWORK_H
+
+
+class AppNetwork {
+
+};
+
+
+#endif //NETWORK_CACHE_APPNETWORK_H
