+++ network-cache/network/announcer/Announcer.cpp
@@ -1,7 +1,3 @@
-//
-// Created by Sergei Alexeev on 11.04.2020.
-//
-
 #include "Announcer.h"
 
 Announcer::Announcer() = default;
