+++ network-cache/network/client/Client.h
@@ -1,0 +1,14 @@
+//
+// Created by Sergei Alexeev on 13.04.2020.
+//
+
+#ifndef NETWORK_CACHE_CLIENT_H
+#define NETWORK_CACHE_CLIENT_H
+
+
+class Client {
+
+};
+
+
+#endif //NETWORK_CACHE_CLIENT_H
