+++ network-cache/network/parser/Parser.h
@@ -1,0 +1,14 @@
+//
+// Created by Sergei Alexeev on 13.04.2020.
+//
+
+#ifndef NETWORK_CACHE_PARSER_H
+#define NETWORK_CACHE_PARSER_H
+
+
+class Parser {
+
+};
+
+
+#endif //NETWORK_CACHE_PARSER_H
